+++ shared/src/error.c
@@ -20,8 +20,8 @@ const char *progname = NULL;
 const char *progvers = NULL;
 
 int exit_status = EXIT_SUCCESS;
-int maximum_errors = 20;
-int number_errors = 0;
+unsigned long maximum_errors = 20;
+unsigned long number_errors = 0;
 
 int crt_line_no = 1;
 const char *crt_file_name = NULL;
